requirements.txt
numpy>=1.24.0
pickle-mixin==1.0.2
streamlit>=1.14.0
streamlit-option-menu==0.3.2
scikit-learn>=1.2.0
setuptools>=58.0.0
click>=7.0,<8.0
protobuf==3.20.3
altair==4.2.2
joblib
openpyxlpandas